requirements.txt
asgiref
autopep8
Django
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
DJANGO
django-import-export
django-clear-cache
django-sb-admin
django-widget-tweaks
psycopg2
django-countries
django-birthday
django-admin-filters
django-admin-autocomplete-filter
django-filter
django-saml2-auth
django-crispy-forms
pandas
django-xhtml2pdf
